Implements basic builds list dashboard
This commit contributes the a working implementation of a build list
dashboard. The application is split into two main parts, the Main modules
and the Builds modules.

The Main modules (or app modules) bootstraps the application and defines
the subscription used to periodically make new requests for data.

The Builds modules defines how a build model looks like and how to render
the builds list and each build. It also defines some logic for messages
and updating the builds part of the app state.

Currently one has to hard-code gitlab url, token, and project id. In the
future this will be replaced with some kind of dialogue box. Should probably
think about the safest way of dealing with storing the token in the browser.

Signed-off-by: Snorre Magnus Davøen <snorremd@gmail.com>